+++ src/components/property/ImageCarousel.tsx
