+++ src/components/property/detail/GrayButton.tsx
