+++ src/components/property/detail/InfoBox.tsx
