+++ src/components/property/detail/section/DealSection.tsx
