+++ src/components/property/detail/section/InfoSection.tsx
